more obj oriented

+++ Battle.py
@@ -1,9 +1,6 @@
 #!/usr/bin/python
-# import appropriate libraries
 import random
 import sqlite3
-# pre-battle setup
-# import DB and player stat table
 # * Creates Battle() object b -> call startBattle with b.startBattle()
 class Battle:
 	playerHp = 100 #placeholder value
@@ -17,14 +14,15 @@ class Battle:
 	]
 	def __init__(self):
 		e = random.randint(0, 4)
-		self.hp = self.enemies[e]['hp']
+		self.hp = int(self.enemies[e]['hp'])
 		self.name = self.enemies[e]['name']
-		self.dmg = self.playerHp * self.enemies[e]['dmg']
+		self.dmg = int(self.playerHp * self.enemies[e]['dmg'])
 	def startBattle(self):
 		print("From the shadows emerges a", self.name)
 		#evantual function for move choice
 		while self.hp > 0 and self.playerHp > 0:
 			#skill table needed
+			# move and damage values are placeholders
 			attack = input("What will you do? (Punch, Block): ").lower()
 			if attack == "punch":
 				self.hp -= 25 #placeholder value
